+++ lib/src/model/product_type.dart
@@ -8,23 +8,23 @@ import '../utils/server_type.dart';
 /// Somehow redundant with [ServerType] and [Flavor].
 enum ProductType implements OffTagged {
   @JsonValue('food')
-  food(offTag: 'food'),
-
+  food(offTag: 'food', flavor: Flavor.openFoodFacts),
   @JsonValue('beauty')
-  beauty(offTag: 'beauty'),
-
+  beauty(offTag: 'beauty', flavor: Flavor.openBeautyFacts),
   @JsonValue('petfood')
-  petFood(offTag: 'petfood'),
-
+  petFood(offTag: 'petfood', flavor: Flavor.openPetFoodFacts),
   @JsonValue('product')
-  product(offTag: 'product');
+  product(offTag: 'product', flavor: Flavor.openProductFacts);
 
   const ProductType({
     required this.offTag,
+    required this.flavor,
   });
 
   @override
   final String offTag;
+
+  final Flavor flavor;
 
   /// Returns the first [ProductType] that matches the [offTag].
   static ProductType? fromOffTag(final String? offTag) =>