+++ lib/src/open_food_api_client.dart
@@ -163,7 +163,7 @@ class OpenFoodAPIClient {
     }
 
     var productUri = uriHelper.getPatchUri(
-      path: '/api/v3/product/$barcode',
+      path: '/api/v3/product/${Uri.encodeComponent(barcode)}',
     );
 
     final Response response = await HttpHelper().doPatchRequest(
@@ -310,7 +310,7 @@ class OpenFoodAPIClient {
     required final bool replaceSubdomain,
   }) {
     final Uri uri = uriHelper.getUri(
-      path: 'product/$barcode',
+      path: 'product/${Uri.encodeComponent(barcode)}',
       addUserAgentParameters: false,
     );
     if (!replaceSubdomain) {
@@ -1173,9 +1173,15 @@ class OpenFoodAPIClient {
   ///   print(orderedNutrients.nutrients[5].name);  // Fiber
   ///   print(orderedNutrients.nutrients[10].name); // Vitamin A
   /// ```
+  ///
+  /// By default, some read-only pseudo nutrients are automatically excluded,
+  /// like `nutrition-score-fr`, as they have negative added-value for the apps:
+  /// the users aren't going to set the values, and they can get the values
+  /// elsewhere.
   static Future<OrderedNutrients> getOrderedNutrients({
     required final OpenFoodFactsCountry country,
     required final OpenFoodFactsLanguage language,
+    final bool excludeReadOnly = true,
     final UriProductHelper uriHelper = uriHelperFoodProd,
   }) async =>
       OrderedNutrients.fromJson(
@@ -1186,6 +1192,7 @@ class OpenFoodAPIClient {
             uriHelper: uriHelper,
           ),
         ),
+        excludeReadOnly: excludeReadOnly,
       );
 
   /// Returns the nutrient hierarchy specific to a country, localized, as JSON